Remove consolidated models/records cmip6 modules; re-export from plugin; update imports; tests+lint green
