+++ src/piddiplatsch/models/__init__.py
@@ -1,5 +1,6 @@
+from piddiplatsch.plugins.cmip6.model import CMIP6DatasetModel as CMIP6DatasetModel
+from piddiplatsch.plugins.cmip6.model import CMIP6FileModel as CMIP6FileModel
+
 from .base import HostingNode as HostingNode
-from .cmip6_model import CMIP6DatasetModel as CMIP6DatasetModel
-from .cmip6_model import CMIP6FileModel as CMIP6FileModel
 
 __all__ = ["CMIP6DatasetModel", "CMIP6FileModel", "HostingNode"]
