--- src/piddiplatsch/models/cmip6_model.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from piddiplatsch.plugins.cmip6.model import CMIP6DatasetModel, CMIP6FileModel
-
-__all__ = ["CMIP6DatasetModel", "CMIP6FileModel"]
