+++ src/piddiplatsch/records/__init__.py
@@ -1,2 +1,2 @@
-from .cmip6_record import CMIP6DatasetRecord as CMIP6DatasetRecord
-from .cmip6_record import CMIP6FileRecord as CMIP6FileRecord
+from piddiplatsch.plugins.cmip6.record import CMIP6DatasetRecord as CMIP6DatasetRecord
+from piddiplatsch.plugins.cmip6.record import CMIP6FileRecord as CMIP6FileRecord
